// typescript/looker/sdk/models.ts
export interface IDictionary<T> {
  [key: string]: T
}
// ...
export interface ICredentialsApi3 {
  /**
   * Unique Id (read-only)
   */
  id?: number
  /**
   * API key client_id (read-only)
   */
  client_id?: string
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsEmail {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * EMail address used for user login
   */
  email?: string
  /**
   * Force the user to change their password upon their next login
   */
  forced_password_reset_at_next_login?: boolean
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * Url with one-time use secret token that the user can use to reset password (read-only)
   */
  password_reset_url?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Link to get this user (read-only)
   */
  user_url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsEmbed {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * Embedder's id for a group to which this user was added during the most recent login (read-only)
   */
  external_group_id?: string
  /**
   * Embedder's unique id for the user (read-only)
   */
  external_user_id?: string
  /**
   * Unique Id (read-only)
   */
  id?: number
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsGoogle {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * Google domain (read-only)
   */
  domain?: string
  /**
   * EMail address (read-only)
   */
  email?: string
  /**
   * Google's Unique ID for this user (read-only)
   */
  google_user_id?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsLDAP {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * EMail address (read-only)
   */
  email?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * LDAP Distinguished name for this user (as-of the last login) (read-only)
   */
  ldap_dn?: string
  /**
   * LDAP Unique ID for this user (read-only)
   */
  ldap_id?: string
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsLookerOpenid {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * EMail address used for user login (read-only)
   */
  email?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * IP address of client for most recent login using credential (read-only)
   */
  logged_in_ip?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Link to get this user (read-only)
   */
  user_url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsOIDC {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * EMail address (read-only)
   */
  email?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * OIDC OP's Unique ID for this user (read-only)
   */
  oidc_user_id?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsSaml {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * EMail address (read-only)
   */
  email?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Timestamp for most recent login using credential (read-only)
   */
  logged_in_at?: string
  /**
   * Saml IdP's Unique ID for this user (read-only)
   */
  saml_user_id?: string
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}

export interface ICredentialsTotp {
  /**
   * Timestamp for the creation of this credential (read-only)
   */
  created_at?: string
  /**
   * Has this credential been disabled? (read-only)
   */
  is_disabled?: boolean
  /**
   * Short name for the type of this kind of credential (read-only)
   */
  type?: string
  /**
   * User has verified (read-only)
   */
  verified?: boolean
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}
// ...
export interface IUser {
  /**
   * URL for the avatar image (may be generic) (read-only)
   */
  avatar_url?: Url
  /**
   * URL for the avatar image (may be generic), does not specify size (read-only)
   */
  avatar_url_without_sizing?: Url
  /**
   * API 3 credentials (read-only)
   */
  credentials_api3?: ICredentialsApi3[]
  credentials_email?: ICredentialsEmail
  /**
   * Embed credentials (read-only)
   */
  credentials_embed?: ICredentialsEmbed[]
  credentials_google?: ICredentialsGoogle
  credentials_ldap?: ICredentialsLDAP
  credentials_looker_openid?: ICredentialsLookerOpenid
  credentials_oidc?: ICredentialsOIDC
  credentials_saml?: ICredentialsSaml
  credentials_totp?: ICredentialsTotp
  /**
   * Full name for display (available only if both first_name and last_name are set) (read-only)
   */
  display_name?: string
  /**
   * EMail address (read-only)
   */
  email?: string
  /**
   * (Embed only) ID of user's group space based on the external_group_id optionally specified during embed user login (read-only)
   */
  embed_group_space_id?: number
  /**
   * First name
   */
  first_name?: string
  /**
   * Array of ids of the groups for this user (read-only)
   */
  group_ids?: number[]
  /**
   * ID string for user's home space
   */
  home_space_id?: string
  /**
   * ID string for user's home folder
   */
  home_folder_id?: string
  /**
   * Unique Id (read-only)
   */
  id?: number
  /**
   * Account has been disabled
   */
  is_disabled?: boolean
  /**
   * Last name
   */
  last_name?: string
  /**
   * User's preferred locale. User locale takes precedence over Looker's system-wide default locale. Locale determines language of display strings and date and numeric formatting in API responses. Locale string must be a 2 letter language code or a combination of language code and region code: 'en' or 'en-US', for example.
   */
  locale?: string
  /**
   * Array of strings representing the Looker versions that this user has used (this only goes back as far as '3.54.0') (read-only)
   */
  looker_versions?: string[]
  /**
   * User's dev workspace has been checked for presence of applicable production projects
   */
  models_dir_validated?: boolean
  /**
   * ID of user's personal space (read-only)
   */
  personal_space_id?: number
  /**
   * ID of user's personal folder (read-only)
   */
  personal_folder_id?: number
  /**
   * User is identified as an employee of Looker (read-only)
   */
  presumed_looker_employee?: boolean
  /**
   * Array of ids of the roles for this user (read-only)
   */
  role_ids?: number[]
  /**
   * Active sessions (read-only)
   */
  sessions?: ISession[]
  /**
   * Per user dictionary of undocumented state information owned by the Looker UI.
   */
  ui_state?: IDictionary<string>
  /**
   * User is identified as an employee of Looker who has been verified via Looker corporate authentication (read-only)
   */
  verified_looker_employee?: boolean
  /**
   * User's roles are managed by an external directory like SAML or LDAP and can not be changed directly. (read-only)
   */
  roles_externally_managed?: boolean
  /**
   * Link to get this item (read-only)
   */
  url?: Url
  /**
   * Operations the current user is able to perform on this object (read-only)
   */
  can?: IDictionary<boolean>
}
// ...
export interface IWriteCredentialsEmail {
  /**
   * EMail address used for user login
   */
  email?: string
  /**
   * Force the user to change their password upon their next login
   */
  forced_password_reset_at_next_login?: boolean
}
// ...
export interface IWriteUser {
  credentials_email?: IWriteCredentialsEmail
  /**
   * First name
   */
  first_name?: string
  /**
   * ID string for user's home space
   */
  home_space_id?: string
  /**
   * ID string for user's home folder
   */
  home_folder_id?: string
  /**
   * Account has been disabled
   */
  is_disabled?: boolean
  /**
   * Last name
   */
  last_name?: string
  /**
   * User's preferred locale. User locale takes precedence over Looker's system-wide default locale. Locale determines language of display strings and date and numeric formatting in API responses. Locale string must be a 2 letter language code or a combination of language code and region code: 'en' or 'en-US', for example.
   */
  locale?: string
  /**
   * User's dev workspace has been checked for presence of applicable production projects
   */
  models_dir_validated?: boolean
  /**
   * Per user dictionary of undocumented state information owned by the Looker UI.
   */
  ui_state?: IDictionary<string>
}